full system test

+++ accountProcessor/src/main/java/com/Account.java
@@ -3,11 +3,11 @@ package accountProcessor.src.main.java.com;
 import java.util.Date;
 
 public class Account {
-    public int accountCode;
+    public long accountCode;
     public double value;
     public Date date;
 
-    public Account(double accountValue, int accountCode, Date date) {
+    public Account(double accountValue, long accountCode, Date date) {
         this.accountCode = accountCode;
         this.date = date;
         this.value = accountValue;
